move hooks to hooks directory
