+++ src/transaction/components/TransactionStatusAction.tsx
@@ -1,5 +1,5 @@
 import { cn, text } from '../../styles/theme';
-import { useGetTransactionStatus } from '../core/useGetTransactionStatus';
+import { useGetTransactionStatus } from '../hooks/useGetTransactionStatus';
 import type { TransactionStatusActionReact } from '../types';
 
 export function TransactionStatusAction({
