+++ src/transaction/components/TransactionToastIcon.tsx
@@ -1,5 +1,5 @@
 import { cn, text } from '../../styles/theme';
-import { useGetTransactionToast } from '../core/useGetTransactionToast';
+import { useGetTransactionToast } from '../hooks/useGetTransactionToast';
 import type { TransactionToastIconReact } from '../types';
 
 export function TransactionToastIcon({ className }: TransactionToastIconReact) {
