+++ src/transaction/hooks/useWriteContracts.ts
@@ -15,7 +15,7 @@ export function useWriteContracts({
       mutation: {
         onError: (e) => {
           if (
-            (e as TransactionExecutionError).cause.name ===
+            (e as TransactionExecutionError)?.cause?.name ===
             'UserRejectedRequestError'
           ) {
             setErrorMessage('User rejected request');
@@ -25,8 +25,6 @@ export function useWriteContracts({
         },
         onSuccess: (id) => {
           setTransactionId(id);
-          // do we need this?
-          // mutation.onSuccess(id);
         },
       },
     });
